refactor(ui): rename BoxMessage to NoItemsMessage
- now it truly reflects its purpose: be a message in a list that has no items
