+++ ui/src/components/NoItemsMessage.vue
